backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
openai
requests
pydantic
aiofiles
python-multipart
